fix: changed a size_t to an int so things wont segfault

+++ ft_printf/Source/object_printing/print_conv_str/float_aux/float_aux.c
@@ -37,7 +37,7 @@ char		*str_add(char *dst, char *src, int n)
 }
 
 char	*str_add_rightside(
-	char *dst, char *src, size_t n, char *overflow)
+	char *dst, char *src, int n, char *overflow)
 {
 	while (n)
 	{
